added entities for data parsing
