+++ Movie Evening/Services/Network/Constants.swift
@@ -10,6 +10,5 @@ import Foundation
 enum Constants {
     static let baseURL = "https://api.kinopoisk.dev/v1.3/"
     static let path = "movie"
-    static let randomPath = "random"
     static let apiKey = "73RGG0S-NPBM7JF-G735BVT-01XY53G"
 }
