+++ Movie Evening/Services/Network/Responses/PreviewFilmResponse.swift
@@ -1,5 +1,5 @@
 //
-//  MovieResponse.swift
+//  PreviewFilmResponse.swift
 //  Movie Evening
 //
 //  Created by Иван Тарасенко on 13.08.2023.
@@ -8,7 +8,7 @@
 import Foundation
 
 // MARK: - Movie Response
-struct MovieResponse: Codable {
+struct PreviewFilmResponse: Codable {
     let docs: [Doc]
 }
 
@@ -22,12 +22,6 @@ struct Doc: Codable {
     let names: [Name]
 }
 
-// MARK: - Name
-struct Name: Codable {
-    let name: String
-    let language: String?
-}
-
 // MARK: - Poster
 struct Poster: Codable {
     let previewURL: String
@@ -36,13 +30,3 @@ struct Poster: Codable {
         case previewURL = "previewUrl"
     }
 }
-
-//// MARK: - Rating
-//struct Rating: Codable {
-//    let imdb: Double
-//    
-//
-//    enum CodingKeys: String, CodingKey {
-//        case imdb
-//    }
-//}
